For fixes

+++ engine/src/move_gen/Generate.rs
@@ -59,7 +59,7 @@ impl MoveGen {
         }
     }
 
-    pub fn calculate_white_pawn_move(&self, position: u64, occupancy: u64, opponent_occupancy: u64) -> u64{
+    pub fn calculate_white_pawn_move(&self, position: usize, occupancy: u64, opponent_occupancy: u64) -> u64{
         let mut result: u64 = 0;
         let forwardPosition = 1 << position - 8;
         let forwardLegal = forwardPosition & occupancy;
